Move crate::schemas to crate::schema in SDK
Module names for Grid features should generally be singular; in this
instance, we talk about "Grid Schema" not "Grid Schemas".
