+++ cli/src/actions/products.rs
@@ -27,7 +27,7 @@ use grid_sdk::protocol::product::payload::{
 use grid_sdk::protocol::product::state::ProductNamespace;
 use grid_sdk::protocol::schema::state::{LatLongBuilder, PropertyValue, PropertyValueBuilder};
 use grid_sdk::protos::IntoProto;
-use grid_sdk::schemas::addressing::GRID_SCHEMA_NAMESPACE;
+use grid_sdk::schema::addressing::GRID_SCHEMA_NAMESPACE;
 use reqwest::Client;
 
 use crate::error::CliError;
