+++ cli/src/actions/schemas.rs
@@ -29,7 +29,7 @@ use grid_sdk::protocol::schema::state::{
     DataType as StateDataType, PropertyDefinition, PropertyDefinitionBuilder,
 };
 use grid_sdk::protos::IntoProto;
-use grid_sdk::schemas::addressing::GRID_SCHEMA_NAMESPACE;
+use grid_sdk::schema::addressing::GRID_SCHEMA_NAMESPACE;
 use reqwest::Client;
 
 use serde::Deserialize;
