+++ contracts/location/src/handler.rs
@@ -438,7 +438,7 @@ mod tests {
             },
         },
         protos::IntoBytes,
-        schemas::addressing::compute_schema_address,
+        schema::addressing::compute_schema_address,
     };
 
     use sawtooth_sdk::processor::handler::{ContextError, TransactionContext};
