+++ contracts/schema/src/state.rs
@@ -19,7 +19,7 @@ use grid_sdk::{
         schema::state::{Schema, SchemaList, SchemaListBuilder},
     },
     protos::{FromBytes, IntoBytes},
-    schemas::addressing::compute_schema_address,
+    schema::addressing::compute_schema_address,
 };
 
 cfg_if! {
