+++ sdk/src/schema/store/error.rs
