+++ sdk/src/store/mod.rs
@@ -39,7 +39,7 @@ pub trait StoreFactory {
     fn get_grid_product_store(&self) -> Box<dyn crate::product::ProductStore>;
     /// Get a new `SchemaStore`
     #[cfg(any(feature = "schema-store-postgres", feature = "schema-store-sqlite"))]
-    fn get_grid_schema_store(&self) -> Box<dyn crate::schemas::SchemaStore>;
+    fn get_grid_schema_store(&self) -> Box<dyn crate::schema::SchemaStore>;
     /// Get a new `TrackAndTraceStore`
     #[cfg(feature = "track-and-trace")]
     fn get_grid_track_and_trace_store(&self)
