+++ sdk/src/store/postgres.rs
@@ -51,8 +51,8 @@ impl StoreFactory for PgStoreFactory {
     }
 
     #[cfg(all(feature = "diesel", feature = "schema-store-postgres"))]
-    fn get_grid_schema_store(&self) -> Box<dyn crate::schemas::SchemaStore> {
-        Box::new(crate::schemas::DieselSchemaStore::new(self.pool.clone()))
+    fn get_grid_schema_store(&self) -> Box<dyn crate::schema::SchemaStore> {
+        Box::new(crate::schema::DieselSchemaStore::new(self.pool.clone()))
     }
 
     #[cfg(feature = "track-and-trace")]
